requirements.txt
flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.0
scholarly>=1.7.0
arxiv>=2.1.0
semantic-kernel>=0.3.0
transformers>=4.34.0
torch>=2.0.0
sentence-transformers>=2.2.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
lxml>=4.9.0
fake-useragent>=1.4.0
retrying>=1.3.0 